+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Uliweb Change Log
 * improve count process
 * Fix Reference and ManyToMany dump and load bug
 * Add `is_in_web()` funciton, so you can test if current frame is in web executation
+* Add `--gevent` support to call command
 
 0.2.6 Version
 -----------------